Factor out writeDump
changelog_begin
changelog_end
